+++ react/FilterNavigatorFlexible.js
@@ -11,8 +11,10 @@ import styles from './searchResult.css'
 const withSearchPageContextProps = Component => ({
   layout,
   initiallyCollapsed,
+  scrollToTop,
   maxItemsDepartment,
   maxItemsCategory,
+  filtersTitleHtmlTag,
 }) => {
   const {
     searchQuery,
@@ -60,8 +62,10 @@ const withSearchPageContextProps = Component => ({
           hiddenFacets={hiddenFacets}
           layout={layout}
           initiallyCollapsed={initiallyCollapsed}
+          scrollToTop={scrollToTop}
           maxItemsDepartment={maxItemsDepartment}
           maxItemsCategory={maxItemsCategory}
+          filtersTitleHtmlTag={filtersTitleHtmlTag}
         />
       </FilterNavigatorContext.Provider>
     </div>