+++ react/components/GalleryRow.js
@@ -1,7 +1,7 @@
 import React, { useRef, useEffect, memo } from 'react'
 import { useInView } from 'react-intersection-observer'
 import { usePixel } from 'vtex.pixel-manager/PixelContext'
-import { useCssHandles } from 'vtex.css-handles'
+import { useCssHandles, applyModifiers } from 'vtex.css-handles'
 import classNames from 'classnames'
 
 import GalleryItem from './GalleryItem'
@@ -20,7 +20,10 @@ const GalleryRow = ({ products, summary, displayMode, itemsPerRow }) => {
       <div
         key={product.productId}
         style={style}
-        className={classNames(handles.galleryItem, 'pa4')}
+        className={classNames(
+          applyModifiers(handles.galleryItem, displayMode),
+          'pa4'
+        )}
       >
         <GalleryItem
           item={product}